backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
sqlalchemy>=2.0.9
alembic>=1.10.3
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
pinecone-client>=2.2.1
neo4j>=5.8.0
langchain>=0.0.200
openai>=0.27.6
python-dotenv>=1.0.0
httpx>=0.24.0
pytest>=7.3.1
tenacity>=8.2.2
redis>=4.5.5
celery>=5.2.7
